Add ability to create twitch clips via GQL
TwitchClipService introduced, originally intended to be a singleton
so it could handle race-conditions given this needed scaled (twitch's
gql endpoint has rate-limits) but honestly the code is so slow it
doesn't matter, doubt this will need scaled anyways.

Clip creation is put right into the existing sync functionality, was
the quickest way to get this working - not the best design though.