Implement validator chain for supported URLs
Majority of this code is copy pasted from the old clipsync, added a
ton of todos based on my thoughts of how to develop this new version.

Plan is to implement some persistence for ensuring requests aren't
handled twice during redeployments, and also maybe reusing it for
caching requests already made - including the created clips.